Convert argument tests to be more unity
